+++ header.php
@@ -1,8 +1,5 @@
 <!DOCTYPE html>
-<!--[if lt IE 7]>      <html class="no-js lt-ie9 lt-ie8 lt-ie7" <?php language_attributes(); ?>> <![endif]-->
-<!--[if IE 7]>         <html class="no-js lt-ie9 lt-ie8" <?php language_attributes(); ?>> <![endif]-->
-<!--[if IE 8]>         <html class="no-js lt-ie9" <?php language_attributes(); ?>> <![endif]-->
-<!--[if gt IE 8]><!--> <html class="no-js" <?php language_attributes(); ?>> <!--<![endif]-->
+<html class="no-js" <?php language_attributes(); ?>>
     <head>
         <meta charset="<?php bloginfo( 'charset' ); ?>">
         <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
@@ -25,11 +22,6 @@
 
     </head>
     <body <?php body_class(); ?>>
-        <!--[if lt IE 7]>
-            <p class="browsehappy">You are using an <strong>outdated</strong> browser.</p>
-            <p class="browsehappy">Please <a href="http://browsehappy.com/">upgrade your browser</a> to improve your experience.</p>
-        <![endif]-->
-        
         <section id="header">
 	        <div class="container">
 		        <div class="row">
@@ -42,7 +34,8 @@
 				        <?php
 							// Call either the responsive_bs_menu() or the non_responsive_bs_menu() here
 							//  If you are using a menu besides 'primary', use this as the first argument in the function.
-							//  You can control whether the menu is right- or left-aligned with $menu_side.  It will be left-aligned unless you set
+							//  You can control whether the menu is right- or left-aligned with $menu_side.
+                            //  It will be left-aligned unless you set
 							//      $menu_side = 'right'
 							//  You can also set up what the text is on the left-hand side of the mobile menu with $brand.
 							//  Lastly, the menu now returns rather than echoing out, so make sure to echo out the menu.