+++ inc/auto-updates.php
@@ -1,0 +1,2 @@
+<?php
+add_filter( 'automatic_updates_is_vcs_checkout', '__return_false', 1 );