+++ template-parts/content-page-home.php
@@ -4,7 +4,7 @@
 		while(have_posts()) {
 			the_post();
 			
-			echo "<article>";
+			echo "<article class='rte'>";
 			the_content();
 			echo "</article>";
 		}